Initial commit
This uses FAISS (Facebook AI Similarity Search) for intent parsing.
This is meant to be used with an LLM, using the speechhandler
plugins to provide context to the LLM, and allowing the LLM to
freestyle. Right now the threshold is set pretty high (50%) on
what intents to use, which differenciates it from the Naomi TTI
plugin, which is set to about 5% and often returns guesses because
if it does not identify an intent, it's only alternative is to
say "I don't understand".